[REF] stock_rest: split services on several files
