+++ stock_rest/__manifest__.py
@@ -6,7 +6,7 @@
     'name': "Stock REST",
     'author': 'NuoBiT Solutions, S.L.',
     'category': 'Stock',
-    'version': '11.0.1.0.1',
+    'version': '11.0.1.0.2',
     'license': 'AGPL-3',
     'website': 'https://github.com/nuobit',
     'depends': [
