+++ stock_rest/services/__init__.py
@@ -1,1 +1,5 @@
 from . import stock_service
+from . import stock_operation_service
+from . import stock_lot_service
+from . import stock_product_service
+
